# server/common/api.yml
openapi: 3.0.1
info:
  title: liver-schedule-manager
  description: The schedule manager for liver
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
...
  /events/{id}:
    get:
      tags:
        - Events
      parameters:
        - name: id
          in: path
          description: The id of the event to retrieve
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Return the event with the specified id
          content: {}
        404:
          description: Example not found
          content: {}
    patch:
      tags:
        - Events
      description: update a new event
      requestBody:
        description: an event
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventBody"
        required: true
      responses:
        200:
          description: Returns inputted events
          content: {}
    delete:
      tags:
        - Events
      description: delete a new event
      parameters:
        - name: id
          in: path
          description: The id of the event to retrieve
          required: true
          schema:
            type: integer
      responses:
        204:
          description: Returns empty body
          content: {}
...
components:
  schemas:
    EventBody:
      title: event
      required:
        - name
      type: object
      properties:
        name:
          type: string
          example: no_stress